20/10/05 solved perfectRect - uno

+++ Week 4/perfectRect/uno.cpp
@@ -9,7 +9,7 @@ long long solution(int w, int h) {
 	long long h_ = h;
 
 	for (int y = 1; y < h; y++) 
-		answer += (double)((w_ * h_) - (w_ * y)) / h_;	
+		answer += (long long)((w_ * h_) - (w_ * y)) / h_;	
 
 	return answer * 2;
 }
@@ -22,11 +22,6 @@ long long solution(int w, int h) {
 	//			break;
 	//	}
 	//}
-
-	////w = 8, h = 12의 원래 직사각형에서는 정사각형 96개 사용가능(8*12)
-	////대각선으로 그으면 총 16개를 사용할 수 없음
-	////정사각형일 때 -> w(h)만큼 사용못함
-	////직사각형일 때 -> min(w,h)*2만큼 사용못함
 
 	//// 1x1 -> 1
